requirements.txt
python-telegram-bot==21.5
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0